+++ packages/twenty-front/src/modules/settings/integrations/components/wrappers/SettingsIntegrationDatabaseConnectionWrapper.tsx
@@ -1,0 +1,74 @@
+import { ReactNode, useEffect } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
+import { IconSettings } from 'twenty-ui';
+
+import { useGetDatabaseConnection } from '@/databases/hooks/useGetDatabaseConnection';
+import { useGetDatabaseConnectionTables } from '@/databases/hooks/useGetDatabaseConnectionTables';
+import { SettingsPageContainer } from '@/settings/components/SettingsPageContainer';
+import { DatabaseConnectionContext } from '@/settings/integrations/contexts/DatabaseConnectionContext';
+import { useSettingsIntegrationCategories } from '@/settings/integrations/hooks/useSettingsIntegrationCategories';
+import { AppPath } from '@/types/AppPath';
+import { SubMenuTopBarContainer } from '@/ui/layout/page/SubMenuTopBarContainer';
+import { useIsFeatureEnabled } from '@/workspace/hooks/useIsFeatureEnabled';
+
+export const SettingsIntegrationDatabaseConnectionWrapper = ({
+  children,
+}: {
+  children: ReactNode;
+}) => {
+  const { databaseKey = '', connectionId = '' } = useParams();
+  const navigate = useNavigate();
+
+  const [integrationCategoryAll] = useSettingsIntegrationCategories();
+  const integration = integrationCategoryAll.integrations.find(
+    ({ from: { key } }) => key === databaseKey,
+  );
+
+  const isAirtableIntegrationEnabled = useIsFeatureEnabled(
+    'IS_AIRTABLE_INTEGRATION_ENABLED',
+  );
+  const isPostgresqlIntegrationEnabled = useIsFeatureEnabled(
+    'IS_POSTGRESQL_INTEGRATION_ENABLED',
+  );
+  const isIntegrationAvailable =
+    !!integration &&
+    ((databaseKey === 'airtable' && isAirtableIntegrationEnabled) ||
+      (databaseKey === 'postgresql' && isPostgresqlIntegrationEnabled));
+
+  const { connection, loading } = useGetDatabaseConnection({
+    databaseKey,
+    connectionId,
+    skip: !isIntegrationAvailable,
+  });
+
+  useEffect(() => {
+    if (!isIntegrationAvailable || (!loading && !connection)) {
+      navigate(AppPath.NotFound);
+    }
+  }, [
+    integration,
+    databaseKey,
+    navigate,
+    isIntegrationAvailable,
+    connection,
+    loading,
+  ]);
+
+  const { tables } = useGetDatabaseConnectionTables({
+    connectionId,
+    skip: !isIntegrationAvailable || !connection,
+  });
+
+  if (!isIntegrationAvailable || !connection) return null;
+  return (
+    <SubMenuTopBarContainer Icon={IconSettings} title="Settings">
+      <SettingsPageContainer>
+        <DatabaseConnectionContext.Provider
+          value={{ connection, integration, databaseKey, tables }}
+        >
+          {children}
+        </DatabaseConnectionContext.Provider>
+      </SettingsPageContainer>
+    </SubMenuTopBarContainer>
+  );
+};
