--- packages/twenty-front/src/pages/settings/integrations/SettingsIntegrationDatabaseConnectionWrapper.tsx
@@ -1,17 +1,0 @@
-import { ReactNode } from 'react';
-import { IconSettings } from 'twenty-ui';
-
-import { SettingsPageContainer } from '@/settings/components/SettingsPageContainer';
-import { SubMenuTopBarContainer } from '@/ui/layout/page/SubMenuTopBarContainer';
-
-export const SettingsIntegrationDatabaseConnectionWrapper = ({
-  children,
-}: {
-  children: ReactNode;
-}) => {
-  return (
-    <SubMenuTopBarContainer Icon={IconSettings} title="Settings">
-      <SettingsPageContainer>{children}</SettingsPageContainer>
-    </SubMenuTopBarContainer>
-  );
-};
